Improve fast log to avoid lookup table

+++ inc/unit/unit.h
@@ -7,7 +7,7 @@
 typedef struct { statements } name; \
 typedef struct { uint32_t fs; settings } name##Tune; \
 name *init_##name(void); \
-double apply_##name(name *p_##unit, const double x); \
+float apply_##name(name *p_##unit, const float x); \
 void tune_##name(name *p_##unit, const name##Tune *p_##tune); \
 void deinit_##name(name *p_##unit);
